+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## [Unreleased]
 
-None
+### BREAKING CHANGES
+
+- All default values of provider elements’ `option-*` attributes are now
+  `undefined`, if you relied on `<intl-displaynames>`’s `option-type` attribute
+  being `language` by default, you need to set it to `language` explicitly now
+
+### Changed
+
+- All `option*` properties are reflective now, meaning if its value is changed,
+  its corresponding attribute will change as well
 
 ---